COVA-3 refactor(pi-server/control-center): split the server entrypoint into multiple routes
